Make the XPC errors regular dictionaries
This was harder than it looked, because I could not include
<xpc/connection.h>. This header declares the variables I need
to define as 'const', and that prevents me from using a
constructor function to initialize them. I kinda need to do this
because the dictionary data structures are so complex.

Also fixed up xpc_release() to avoid freeing a statically allocated
object (such as our errors). This would still have been a problem
with our old error implementation.